requirements.txt
# API and web server
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-multipart>=0.0.5
jinja2>=3.1.3

# Audio/Video processing
openai>=1.17.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
python-magic>=0.4.27

# Database
supabase>=0.7.1
asyncpg>=0.25.0
psycopg2-binary>=2.9.9

# Utils
python-dotenv>=0.19.1
pandas>=2.2.1
numpy>=1.26.4
aiohttp>=3.8.1
tenacity>=8.2.3

# LinkedIn scraping
scrapy>=2.11.0
requests>=2.26.0

transformers==4.31.0
torch>=2.0.1
SQLAlchemy>=1.4,<2.0
httpx>=0.20.0
beautifulsoup4==4.12.2
scrapy-rotating-proxies==0.6.2

# Date and time handling
pytz>=2021.3
python-dateutil>=2.8.2

# Type checking
mypy>=0.910
types-requests>=2.26.0

# For BrightData integration
uuid>=1.30
urllib3>=1.26.7 